+++ apps/services/your-id-service/src/app.ts
@@ -1,0 +1,16 @@
+import express from 'express';
+import cors from 'cors';
+
+import userRouter from './routes/userRoutes'
+
+// Express app creation
+const app = express();
+
+// Middlewares
+app.use(express.json());
+app.use(cors());
+
+// Routes
+app.use('/user', userRouter);
+
+export { app };
